buttons

+++ stage2/macro.h
@@ -2,7 +2,7 @@
 #define _MACRO_H_
 
 typedef unsigned char u8;
-typedef unsigned int u16;
+typedef unsigned short int u16;
 
 #define F_CPU 8000000UL
